Add automatic admin creation and session
